Write motor position i.o.t. read upon next power-up
Motor position (variable 'count') is stored onto EEPROM
so that the position of the motor can be read on the next power-up,
in order to realign the motor to the reference position.

e.g. if motor did 1/4 of a turn before it was powered off,
then count=128 is saved to memory (EEPROM), and at next power-up
ref-count=512-128=384, so the motor will need to advanced 384 steps
in order to reach the reference position.